v0.2

+++ Bitcask/Bitcask/Bitcask/MemdirEntry.cs
@@ -14,12 +14,12 @@ namespace Bitcask
         /// Value_Position: The Position of the Value in the Binary File
         /// TimeStamp: TimeStamp
         /// </summary>
-        public long FileID { get; set; }
+        public ulong FileID { get; set; }
         public int Value_Size { get; set; }
         public int Value_Position { get; set; }
         public DateTime TimeStamp { get; set; } = DateTime.Now;
 
-        public MemdirEntry(int fieldid, int value_size, int value_position)
+        public MemdirEntry(ulong fieldid, int value_size, int value_position)
         {
             FileID = fieldid;
             Value_Size = value_size;
